requirements.txt
## The following requirements were added by pip freeze:
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-filter==22.1
django-import-export==3.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
Markdown==3.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2==2.9.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.6
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0